+++ dpp-backend/digitalproductpass/src/main/java/org/eclipse/tractusx/digitalproductpass/http/controllers/api/ApiController.java
@@ -25,7 +25,6 @@
 
 package org.eclipse.tractusx.digitalproductpass.http.controllers.api;
 
-import com.fasterxml.jackson.databind.JsonNode;
 import io.swagger.v3.oas.annotations.Hidden;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
@@ -44,16 +43,19 @@ import org.eclipse.tractusx.digitalproductpass.models.http.requests.DiscoverySea
 import org.eclipse.tractusx.digitalproductpass.models.http.requests.Search;
 import org.eclipse.tractusx.digitalproductpass.models.http.requests.SingleApiRequest;
 import org.eclipse.tractusx.digitalproductpass.models.http.requests.TokenRequest;
-import org.eclipse.tractusx.digitalproductpass.models.manager.Process;
 import org.eclipse.tractusx.digitalproductpass.models.manager.Status;
 import org.eclipse.tractusx.digitalproductpass.models.negotiation.Dataset;
 import org.eclipse.tractusx.digitalproductpass.models.passports.PassportResponse;
 import org.eclipse.tractusx.digitalproductpass.services.AasService;
 import org.eclipse.tractusx.digitalproductpass.services.AuthenticationService;
+import org.eclipse.tractusx.digitalproductpass.services.ContractService;
 import org.eclipse.tractusx.digitalproductpass.services.DataTransferService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.env.Environment;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RestController;
 import utils.HttpUtil;
 import utils.JsonUtil;
 import utils.LogUtil;
@@ -79,13 +81,13 @@ public class ApiController {
     private @Autowired HttpServletResponse httpResponse;
     private @Autowired DataTransferService dataService;
     private @Autowired AasService aasService;
+    private @Autowired ContractService contractService;
     private @Autowired Environment env;
     private @Autowired AuthenticationService authService;
     private @Autowired PassportConfig passportConfig;
     private @Autowired HttpUtil httpUtil;
     private @Autowired JsonUtil jsonUtil;
     private @Autowired ProcessManager processManager;
-    private @Autowired ContractController contractController;
     private @Autowired SingleApiConfig singleApiConfig;
 
     /** METHODS **/
@@ -122,97 +124,13 @@ public class ApiController {
             return httpUtil.buildResponse(response, httpResponse);
         }
         try {
-            return callGetData(tokenRequestBody);
+            return contractService.getDataCall(httpRequest, httpResponse, tokenRequestBody);
         } catch (Exception e) {
             response.message = e.getMessage();
             return httpUtil.buildResponse(response, httpResponse);
         }
     }
-    private Response callGetData(TokenRequest tokenRequestBody){
-        Response response = httpUtil.getInternalError();
-        // Check for the mandatory fields
-        List<String> mandatoryParams = List.of("processId", "contractId", "token");
-        if (!jsonUtil.checkJsonKeys(tokenRequestBody, mandatoryParams, ".", false)) {
-            response = httpUtil.getBadRequest("One or all the mandatory parameters " + mandatoryParams + " are missing");
-            return httpUtil.buildResponse(response, httpResponse);
-        }
 
-        // Check for processId
-        String processId = tokenRequestBody.getProcessId();
-        if (!processManager.checkProcess(httpRequest, processId)) {
-            response = httpUtil.getBadRequest("The process id does not exists!");
-            return httpUtil.buildResponse(response, httpResponse);
-        }
-
-
-        Process process = processManager.getProcess(httpRequest, processId);
-        if (process == null) {
-            response = httpUtil.getBadRequest("The process id does not exists!");
-            return httpUtil.buildResponse(response, httpResponse);
-        }
-
-        // Get status to check for contract id
-        String contractId = tokenRequestBody.getContractId();
-        Status status = processManager.getStatus(processId);
-
-        if (status.historyExists("contract-decline")) {
-            response = httpUtil.getForbiddenResponse("The contract for this passport has been declined!");
-            return httpUtil.buildResponse(response, httpResponse);
-        }
-        if (status.historyExists("negotiation-canceled")) {
-            response = httpUtil.getForbiddenResponse("This negotiation has been canceled! Please request a new one");
-            return httpUtil.buildResponse(response, httpResponse);
-        }
-
-        // Check if the contract id is correct
-        Map<String, Dataset> availableContracts = processManager.loadDatasets(processId);
-        String seedId = String.join("|",availableContracts.keySet()); // Generate Seed
-        // Check the validity of the token
-        String expectedToken = processManager.generateToken(process, seedId);
-        String token = tokenRequestBody.getToken();
-        if (!expectedToken.equals(token)) {
-            response = httpUtil.getForbiddenResponse("The token is invalid!");
-            return httpUtil.buildResponse(response, httpResponse);
-        }
-        Dataset dataset = availableContracts.get(contractId);
-
-        if (dataset == null) {
-            response.message = "The Contract Selected was not found!";
-            return httpUtil.buildResponse(response, httpResponse);
-        }
-
-        if (!status.historyExists("data-received")) {
-            status = processManager.getStatus(processId); // Retry to get the status before giving an error
-            if(!status.historyExists("data-received")) {
-                response = httpUtil.getNotFound("The data is not available!");
-                return httpUtil.buildResponse(response, httpResponse);
-            }
-        }
-
-        if (status.historyExists("data-retrieved")) {
-            response = httpUtil.getNotFound("The data was already retrieved and is no longer available!");
-            return httpUtil.buildResponse(response, httpResponse);
-        }
-        String semanticId = status.getSemanticId();
-        JsonNode passport = processManager.loadPassport(processId);
-        if (passport == null) {
-            response = httpUtil.getNotFound("Failed to load passport!");
-            return httpUtil.buildResponse(response, httpResponse);
-        }
-        Map<String, Object> negotiation = processManager.loadNegotiation(processId);
-        Map<String, Object> transfer = processManager.loadTransfer(processId);
-        response = httpUtil.getResponse();
-        response.data = Map.of(
-                "metadata", Map.of(
-                        "contract", dataset,
-                        "negotiation", negotiation,
-                        "transfer", transfer
-                ),
-                "aspect", passport,
-                "semanticId", semanticId
-        );
-        return httpUtil.buildResponse(response, httpResponse);
-    }
     /**
      * HTTP POST method to retrieve the Passport with an API Key authentication.
      * <p>
@@ -246,7 +164,7 @@ public class ApiController {
             discoverySearch.setId(singleApiRequestBody.getDiscoveryId());
             discoverySearch.setType(singleApiRequestBody.getDiscoveryIdType());
             //Call Create function
-            Response createResponse = contractController.createCall(discoverySearch);
+            Response createResponse = contractService.createCall(httpResponse, discoverySearch);
             //The Status from the response must 200 to proceed
             if (createResponse.getStatus() != 200) {
                 return createResponse;
@@ -267,7 +185,7 @@ public class ApiController {
             searchBody.setProcessId(createResponseData.get("processId"));
             searchBody.setChildren(singleApiRequestBody.getChildren());
             searchBody.setSemanticId(singleApiRequestBody.getSemanticId());
-            Response searchResponse = contractController.searchCall(searchBody);
+            Response searchResponse = contractService.searchCall(httpRequest, httpResponse, searchBody);
             //The Status from the response must 200 to proceed
             if (searchResponse.getStatus() != 200) {
                 return searchResponse;
@@ -289,7 +207,7 @@ public class ApiController {
             String contractId = contracts.entrySet().stream().findFirst().get().getKey();
             System.out.println("CONTRACTID: "+ contractId);
             tokenRequest.setContractId(contractId);
-            Response agreeResponse = contractController.agreeCall(tokenRequest);
+            Response agreeResponse = contractService.doContractAgreement(httpRequest, httpResponse, tokenRequest);
             LogUtil.printMessage("[SINGLE API] [PROCESS "+processId + "] Agreed with a contract and started the contract negotiation!");
             //The Status from the response must 200 to proceed
             if (agreeResponse.getStatus() != 200) {
@@ -309,7 +227,7 @@ public class ApiController {
                 if (status.historyExists("transfer-completed") || status.historyExists("data-received")) {
                     break;
                 }
-                Response statusResponse = contractController.statusCall(searchBody.getProcessId());
+                Response statusResponse = contractService.statusCall(httpResponse, searchBody.getProcessId());
                 //The Status from the response must 200 to proceed
                 if (statusResponse.getStatus() != 200) {
                     return statusResponse;
@@ -329,7 +247,7 @@ public class ApiController {
             }
             LogUtil.printMessage("[SINGLE API] [PROCESS "+processId + "] Transfer process completed! Retrieving the Passport Data!");
             //Call getData function
-            response = callGetData(tokenRequest);
+            response = contractService.getDataCall(httpRequest, httpResponse, tokenRequest);
 
             return response;
         } catch (Exception e) {